Restructure the packages
